+++ core/interfaces.go
@@ -154,6 +154,7 @@ type StorageAdder interface {
 	FinalizeOrder(ctx context.Context, order *corepb.Order) error
 	AddPendingAuthorizations(ctx context.Context, req *sapb.AddPendingAuthorizationsRequest) (*sapb.AuthorizationIDs, error)
 	SetOrderError(ctx context.Context, order *corepb.Order) error
+	UpdateCertificateStatus(ctx context.Context, status *sapb.CertificateStatus) error
 }
 
 // StorageAuthority interface represents a simple key/value
